### 2017-10-03 #423 @cxhernandez
>@cxhernandez if you have any suggestions for potential reviewers then please mention them here in this thread.

@anotherjoshsmith was very helpful in his review of https://github.com/openjournals/joss-reviews/issues/188
