### 2017-10-04 #423 @anotherjoshsmith
@arfon @arokem @cxhernandez I'd be happy to review MDEntropy
