### 2017-10-08 #427 @whedon
[REVIEW]: MDEntropy: Information-Theoretic Analyses for Molecular Dynamics
**Submitting author:** @cxhernandez (<a href="http://orcid.org/0000-0002-8146-5904">Carlos Xavier Hernández</a>)
**Repository:** <a href="https://github.com/msmbuilder/mdentropy/" target ="_blank">https://github.com/msmbuilder/mdentropy/</a>
**Version:** v0.3.0
