### 2017-10-20 msmbuilder/mdentropy#65 @anotherjoshsmith
[ JOSS Review ]
I've played with MDEntropy over the last couple days and haven't run into any problems with the core functionality. The only box left [unchecked](https://github.com/openjournals/joss-reviews/issues/427) is due to an installation issue with a fresh Python 2.7.14 environment. Looks like MDEntropy asks for Python 3.5* on the py27 channel.

- [ ] Fix Python 2.7 install

As soon as this is checked, I'll suggest JOSS accept your submission!

A couple of other suggestions, just to help out potential first time users:
- [x] link to [MDEntropy Website]( http://msmbuilder.org/mdentropy) from the GitHub README. I only found the website after looking reading the paper.
- [x] note specific dependencies for the Mutual Information example [here](http://msmbuilder.org/mdentropy/0.3.0/examples/index.html). Should just be matplotlib, nglview, and msmexplorer.

Looks like another useful tool for MD analysis, nice work!
