### 2017-10-20 #427 @anotherjoshsmith
@cxhernandez I've opened [an issue](https://github.com/msmbuilder/mdentropy/issues/65) for the review. I've had no trouble with the functionality, and the documentation on the website was thorough enough for me to successfully test several of the convenience functions, information calculations, and feature metrics.

@arokem As soon as the installation issue for Python 2.7 is fixed, I think MDEntropy is ready for acceptance.
