### 2017-11-01 msmbuilder/mdentropy#65 @cxhernandez
> The only box left unchecked is due to an installation issue with a fresh Python 2.7.14 environment. Looks like MDEntropy asks for Python 3.5* on the py27 channel.

I actually need to update the documentation! The package only supports python 3.4+ currently, due to some design choices.

The other points, I do right now! Thanks!
