### 2017-11-01 c66d942
Update installation.rst
