### 2017-11-01 c73a6d5
Update index.rst
