### 2017-11-14 #427 @arokem
Hi - checking in here again. I think it's fine for the software to only support Python 3, as long as this is clearly documented. Is that the case here?
