### 2017-11-14 #427 @anotherjoshsmith
Thanks, @arokem. With the specification of Python 3 support and linking to documentation from the Github README, MDEntropy is ready for acceptance!
