### 2017-11-14 msmbuilder/mdentropy#65 @anotherjoshsmith
Thanks @cxhernandez. With those updates, I think MDEntropy is ready for acceptance by JOSS.
