### 2017-11-14 #427 @arokem
@arfon : I believe this paper can now be published.
