### 2017-11-15 #427 @arfon
@anotherjoshsmith - many thanks for your review here and to @arokem for editing this one ✨

@cxhernandez - your paper is now accepted into JOSS and your DOI is https://doi.org/10.21105/joss.00427 ⚡️ 🚀 :boom:
